Fix integer overflow in IndexInputStream.Read(), #1158
When input.Length - input.Position exceeds int.MaxValue, casting to int
causes overflow resulting in negative values. This fix uses long for the
remaining calculation before safely casting to int via Math.Min.

Added unit test TestGitHubIssue1158_IndexInputStream_Read_IntegerOverflow
to demonstrate and verify the fix.

+++ src/Lucene.Net.Replicator/IndexInputInputStream.cs
@@ -65,8 +65,8 @@ namespace Lucene.Net.Replicator
 
         public override int Read(byte[] buffer, int offset, int count)
         {
-            int remaining = (int)(input.Length - input.Position); // LUCENENET specific: Renamed from getFilePointer() to match FileStream
-            int readCount = Math.Min(remaining, count);
+            long remaining = input.Length - input.Position; // LUCENENET specific: Renamed from getFilePointer() to match FileStream
+            int readCount = (int)Math.Min(remaining, count);
             input.ReadBytes(buffer, offset, readCount);
             return readCount;
         }
